Avoid recursion in replacing socket with GuardedSocket
In a project with a lot of tests, a recursion error is eventually raised due
to the socket being replaced with a GuardedSocket at the start of each test,
but this is not undone after the test.

Add a check and if the socket has already been replaced, skip doing it again.